[backend] refactor: use application layer for several user mutations
